V1.6.7.0 Significant extension to the MegaCopy functionality: When an area has been copied the font pixels can now be - Shifted left/right/up/down - Mirrored horizontally or vertically - Inverted - If the area is square the pixels can be rotated left or right - If the area contains unique characters then the modified font area can be pasted directly into the source characters (into any of the 4 fonts)
Using the mouse wheel to select the next/prev character has been extended with the control key. When pressing CTRL and using the mouse wheel the character selector will move a row at a time (jump 32 characters).
This allows for quick left-right and up-down movement in the font selection area.

The characters in a font can also be shifted left-right with or without creating a hole at the current selected location